refactored campgrounds, comments and index routes
