+++ app.js
@@ -1,5 +1,3 @@
-const { serializeUser } = require('passport');
-
 var express = require('express'),
     app = express(),
     bodyParser = require("body-parser"),
@@ -10,6 +8,11 @@ var express = require('express'),
     Comment = require("./models/comment"),
     User = require("./models/user"),
     seedDB = require("./seeds");
+
+// requiring routes
+var commentRoutes = require("./routes/comments"),
+    campgroundRoutes = require("./routes/campgrounds"),
+    indexRoutes = require("./routes/index");
 
 mongoose.set('useUnifiedTopology', true);
 mongoose.connect("mongodb://localhost/yelp_camp",{useNewUrlParser:true});
@@ -37,148 +40,10 @@ app.use(function(req,res,next){
     next();
 });
 
+app.use("/",indexRoutes);
+app.use("/campgrounds",campgroundRoutes);
+app.use("/campgrounds/:id/comments",commentRoutes);
 
-// ROUTES ====================
-
-app.get("/", function(req,res){
-    res.render("landing");
-});
-
-//INDEX : Show all camgrounds from DB
-app.get("/campgrounds",function(req,res){
-    //GET ALL CG from DB
-    Campground.find({},function (err,allCampgrounds) {
-        if(err){
-            console.log(err);
-        } else{
-            res.render("campgrounds/index",{campgrounds:allCampgrounds});
-        }
-    });
-    
-});
-
-//CREATE ROUTE : add a new campground to DB
-app.post("/campgrounds",function (req,res) {
-   //get data from form and add to campgrounds array
-   var name = req.body.name;
-   var image = req.body.image;
-   var desc = req.body.description;
-   var newCampground = {name : name, image : image, description : desc};
-   //create a new CG and save to DB
-   Campground.create(newCampground,function(err,newlyCreated){
-        if(err){
-            console.log(err);
-        } else{
-             //redirect back to campgrounds page 
-             res.redirect("/campgrounds");
-        }
-   });
-  
-});
-
-//NEW :  Displays form to add a new CG to DB
-app.get("/campgrounds/new",function(req,res){
-    res.render("campgrounds/new");
-});
-
-//SHOW : Displays more data about particular CG from DB (should be after every NEW route)
-app.get("/campgrounds/:id",function(req,res){
-    //Find CG with provided ID
-    Campground.findById(req.params.id).populate("comments").exec(function(err,foundCampground){
-        if(err){
-            console.log(err);
-        } else{
-            //render show template with that campground
-                res.render("campgrounds/show",{campground:foundCampground});
-        }
-    });
-});
-
-// =============
-//COMMENTS ROUTES
-//==============
-
-app.get("/campgrounds/:id/comments/new",isLoggedIn,function(req,res){
-    //find campground by id
-    Campground.findById(req.params.id,function(err,campground){
-        if(err){
-            console.log(err);
-        } else {
-            res.render("comments/new",{campground:campground});
-        }
-    });
-});
-
-app.post("/campgrounds/:id/comments",isLoggedIn,function(req,res){
-    //look up campgrounds using ID
-    Campground.findById(req.params.id, function(err,campground){
-        if(err){
-            console.log(err);
-            res.redirect("/campgrounds");
-        } else{
-             //create new comment
-             Comment.create(req.body.comment,function(err,comment){
-                if(err){
-                    console.log(err);
-                } else{
-                    //connect new comment to campground
-                    campground.comments.push(comment);
-                    campground.save();
-                    //redirect to campground showpage
-                    res.redirect('/campgrounds/' + campground._id);
-                }
-             });           
-        }
-    });
-});
-
-//=========
-// AUTH ROUTES
-//=========
-
-//show register form
-app.get("/register",function(req,res){
-    res.render("register");
-});
-
-//handle signup logic
-app.post("/register",function(req,res){
-    var newUser = new User({username:req.body.username});
-    User.register(newUser,req.body.password,function(err,user){
-        if(err){
-            console.log(err);
-            return res.render("register");
-        }
-        passport.authenticate("local")(req,res,function(){
-            res.redirect("/campgrounds");
-        });
-    })
-});
-
-//show login form
-app.get("/login",function(req,res){
-    res.render("login");
-});
-
-//handle login logic
-app.post("/login",passport.authenticate("local",
-    {
-    successRedirect:"/campgrounds",
-    failureRedirect:"/login"
-    }),function(req,res){});
-
-//logout route
-app.get("/logout",function(req,res){
-    req.logout();
-    res.redirect("/campgrounds");
-});
-
-function isLoggedIn(req,res,next){
-    if(req.isAuthenticated()){
-        return next();
-    }
-    res.redirect("/login");
-}
 
 app.listen(5000, function () {
     console.log("Server has started");
